renamed messages to promo-messages
