+++ bootstrap.scss
@@ -16,4 +16,4 @@
 @import 'n-cookie-message/main';
 @import 'n-interactive-tour/main';
 @import 'message-prompts/main';
-@import 'messages/main';
+@import 'promo-messages/main';
